ai-service/requirements.txt
# AI Service - Reality Firewall Forensic Detection Backend
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Scientific Computing
numpy>=1.26.0
scipy>=1.14.0
scikit-learn>=1.5.0

# Image Processing
Pillow>=10.4.0
opencv-python-headless>=4.10.0

# Audio Processing
librosa>=0.10.2
soundfile>=0.12.1

# Face Detection
facenet-pytorch>=2.6.0

# Deep Learning (Phase 2)
torch>=2.4.0
torchvision>=0.19.0
timm>=1.0.0

# EXIF Parsing
exifread>=3.0.0

# Utilities
filetype>=1.2.0
pydantic>=2.9.0
